Remove parameters that sole instances never need
