+++ spynnaker/pyNN/models/neuron/synaptic_manager.py
@@ -212,10 +212,12 @@ class SynapticManager(object):
         # 4 for the size of the direct addresses matrix in bytes
         return 2 * BYTES_PER_WORD
 
-    def _get_max_row_info(
+    def __get_max_row_info(
             self, synapse_info, post_vertex_slice, app_edge,
             machine_time_step):
         """ Get the maximum size of each row for a given slice of the vertex
+
+        :rtype: MaxRowInfo
         """
         key = (synapse_info, post_vertex_slice.lo_atom,
                post_vertex_slice.hi_atom)
@@ -241,7 +243,7 @@ class SynapticManager(object):
 
     def __add_synapse_size(self, memory_size, synapse_info, post_vertex_slice,
                            in_edge, machine_time_step):
-        max_row_info = self._get_max_row_info(
+        max_row_info = self.__get_max_row_info(
             synapse_info, post_vertex_slice, in_edge, machine_time_step)
         n_atoms = in_edge.pre_vertex.n_atoms
         memory_size = self.__poptable_type.get_next_allowed_address(
@@ -324,8 +326,7 @@ class SynapticManager(object):
                 vertex_slice, application_graph, app_vertex) +
             self._get_synaptic_blocks_size(
                 vertex_slice, in_edges, machine_time_step) +
-            self.__poptable_type.get_master_population_table_size(
-                vertex_slice, in_edges) +
+            self.__poptable_type.get_master_population_table_size(in_edges) +
             self._get_size_of_generator_information(in_edges))
 
     def _reserve_memory_regions(
@@ -647,8 +648,7 @@ class SynapticManager(object):
                             pre_slices, pre_slice_index, post_slices,
                             post_slice_index, pre_vertex_slice,
                             post_vertex_slice, app_edge,
-                            self.__n_synapse_types,
-                            single_synapses, master_pop_table_region,
+                            self.__n_synapse_types, single_synapses,
                             weight_scales, machine_time_step, rinfo,
                             all_syn_block_sz, block_addr, single_addr,
                             machine_edge=machine_edge)
@@ -663,10 +663,8 @@ class SynapticManager(object):
             (synapse_info, pre_slices, pre_vertex_slice, pre_slice_index,
                 app_edge, rinfo) = gen_data
             block_addr, index = self.__generate_on_chip_data(
-                spec, synapse_info,
-                pre_slices, pre_slice_index, post_slices,
-                post_slice_index, pre_vertex_slice,
-                post_vertex_slice, master_pop_table_region, rinfo,
+                synapse_info, pre_slices, pre_slice_index, post_slices,
+                post_slice_index, pre_vertex_slice, post_vertex_slice, rinfo,
                 all_syn_block_sz, block_addr, machine_time_step, app_edge,
                 generator_data)
             key = (synapse_info, pre_vertex_slice.lo_atom,
@@ -696,16 +694,15 @@ class SynapticManager(object):
         return generator_data
 
     def __generate_on_chip_data(
-            self, spec, synapse_info, pre_slices,
-            pre_slice_index, post_slices, post_slice_index, pre_vertex_slice,
-            post_vertex_slice, master_pop_table_region, rinfo,
-            all_syn_block_sz, block_addr, machine_time_step,
-            app_edge, generator_data):
+            self, synapse_info, pre_slices, pre_slice_index, post_slices,
+            post_slice_index, pre_vertex_slice, post_vertex_slice, rinfo,
+            all_syn_block_sz, block_addr, machine_time_step, app_edge,
+            generator_data):
         """ Generate data for the synapse expander
         """
 
         # Get the size of the matrices that will be required
-        max_row_info = self._get_max_row_info(
+        max_row_info = self.__get_max_row_info(
             synapse_info, post_vertex_slice, app_edge, machine_time_step)
 
         # If delay edge exists, tell this about the data too, so it can
@@ -730,9 +727,8 @@ class SynapticManager(object):
             synaptic_matrix_offset = \
                 self.__poptable_type.get_next_allowed_address(block_addr)
             index = self.__poptable_type.update_master_population_table(
-                spec, synaptic_matrix_offset,
-                max_row_info.undelayed_max_words,
-                rinfo.first_key_and_mask, master_pop_table_region)
+                synaptic_matrix_offset, max_row_info.undelayed_max_words,
+                rinfo.first_key_and_mask)
             n_bytes_undelayed = (
                 max_row_info.undelayed_max_bytes * pre_vertex_slice.n_atoms)
             block_addr = synaptic_matrix_offset + n_bytes_undelayed
@@ -741,7 +737,7 @@ class SynapticManager(object):
             synaptic_matrix_offset //= BYTES_PER_WORD
         elif rinfo is not None:
             index = self.__poptable_type.update_master_population_table(
-                spec, 0, 0, rinfo.first_key_and_mask, master_pop_table_region)
+                0, 0, rinfo.first_key_and_mask)
 
         if block_addr > all_syn_block_sz:
             raise Exception(
@@ -763,9 +759,8 @@ class SynapticManager(object):
                 self.__poptable_type.get_next_allowed_address(
                     block_addr)
             d_index = self.__poptable_type.update_master_population_table(
-                spec, delayed_synaptic_matrix_offset,
-                max_row_info.delayed_max_words,
-                delay_rinfo.first_key_and_mask, master_pop_table_region)
+                delayed_synaptic_matrix_offset, max_row_info.delayed_max_words,
+                delay_rinfo.first_key_and_mask)
             n_bytes_delayed = (
                 max_row_info.delayed_max_bytes * pre_vertex_slice.n_atoms *
                 n_delay_stages)
@@ -775,8 +770,7 @@ class SynapticManager(object):
             delayed_synaptic_matrix_offset //= BYTES_PER_WORD
         elif delay_rinfo is not None:
             d_index = self.__poptable_type.update_master_population_table(
-                spec, 0, 0, delay_rinfo.first_key_and_mask,
-                master_pop_table_region)
+                0, 0, delay_rinfo.first_key_and_mask)
 
         if block_addr > all_syn_block_sz:
             raise Exception(
@@ -805,9 +799,8 @@ class SynapticManager(object):
             self, spec, synaptic_matrix_region, synapse_info, pre_slices,
             pre_slice_index, post_slices, post_slice_index, pre_vertex_slice,
             post_vertex_slice, app_edge, n_synapse_types, single_synapses,
-            master_pop_table_region, weight_scales, machine_time_step,
-            rinfo, all_syn_block_sz, block_addr, single_addr,
-            machine_edge):
+            weight_scales, machine_time_step, rinfo, all_syn_block_sz,
+            block_addr, single_addr, machine_edge):
         (row_data, row_length, delayed_row_data, delayed_row_length,
          delayed_source_ids, delay_stages) = self.__synapse_io.get_synapses(
              synapse_info, pre_slices, pre_slice_index, post_slices,
@@ -839,12 +832,11 @@ class SynapticManager(object):
             block_addr, single_addr, index = self.__write_row_data(
                 spec, synapse_info.connector, pre_vertex_slice,
                 post_vertex_slice, row_length, row_data, rinfo,
-                single_synapses, master_pop_table_region,
-                synaptic_matrix_region, block_addr, single_addr, app_edge,
-                synapse_info)
+                single_synapses, synaptic_matrix_region, block_addr,
+                single_addr, app_edge, synapse_info)
         elif rinfo is not None:
             index = self.__poptable_type.update_master_population_table(
-                spec, 0, 0, rinfo.first_key_and_mask, master_pop_table_region)
+                0, 0, rinfo.first_key_and_mask)
         del row_data
 
         if block_addr > all_syn_block_sz:
@@ -862,13 +854,11 @@ class SynapticManager(object):
             block_addr, single_addr, d_index = self.__write_row_data(
                 spec, synapse_info.connector, pre_vertex_slice,
                 post_vertex_slice, delayed_row_length, delayed_row_data,
-                delay_rinfo, single_synapses, master_pop_table_region,
-                synaptic_matrix_region, block_addr, single_addr, app_edge,
-                synapse_info)
+                delay_rinfo, single_synapses, synaptic_matrix_region,
+                block_addr, single_addr, app_edge, synapse_info)
         elif delay_rinfo is not None:
             d_index = self.__poptable_type.update_master_population_table(
-                spec, 0, 0, delay_rinfo.first_key_and_mask,
-                master_pop_table_region)
+                0, 0, delay_rinfo.first_key_and_mask)
         del delayed_row_data
 
         if block_addr > all_syn_block_sz:
@@ -898,16 +888,15 @@ class SynapticManager(object):
     def __write_row_data(
             self, spec, connector, pre_vertex_slice, post_vertex_slice,
             row_length, row_data, rinfo, single_synapses,
-            master_pop_table_region, synaptic_matrix_region,
-            block_addr, single_addr, app_edge, synapse_info):
+            synaptic_matrix_region, block_addr, single_addr, app_edge,
+            synapse_info):
         if row_length == 1 and self.__is_direct(
                 single_addr, connector, pre_vertex_slice, post_vertex_slice,
                 app_edge, synapse_info):
             single_rows = row_data.reshape(-1, 4)[:, 3]
             single_synapses.append(single_rows)
             index = self.__poptable_type.update_master_population_table(
-                spec, single_addr, 1, rinfo.first_key_and_mask,
-                master_pop_table_region, is_single=True)
+                single_addr, 1, rinfo.first_key_and_mask, is_single=True)
             single_addr += len(single_rows) * BYTES_PER_WORD
         else:
             block_addr = self._write_padding(
@@ -915,8 +904,7 @@ class SynapticManager(object):
             spec.switch_write_focus(synaptic_matrix_region)
             spec.write_array(row_data)
             index = self.__poptable_type.update_master_population_table(
-                spec, block_addr, row_length,
-                rinfo.first_key_and_mask, master_pop_table_region)
+                block_addr, row_length, rinfo.first_key_and_mask)
             block_addr += len(row_data) * BYTES_PER_WORD
         return block_addr, single_addr, index
